Implemented a simplified markdown parser for notes
The actual markdown parser was suprisingly simple but adding
support for [title](url), ![title](url) and [^2] i.e.
links, images, and footnotes required a bit more code.

This should be expandable to allow us to have more semantic
information in our notes (i.e ^^pick: http://github.com/gg to denote
a pick of the week, so we can parse them all out and display
them in their own section of the website